+++ packages/atproto/lib/src/lex_annotations.g.dart
@@ -276,6 +276,14 @@ const comAtprotoServerDefsInviteCodeUse =
 const comAtprotoServerGetAccountInviteCodes =
     LexObject(ids.comAtprotoServerGetAccountInviteCodes);
 
+/// `com.atproto.temp.requestPhoneVerification#main`
+const comAtprotoTempRequestPhoneVerification =
+    LexObject(ids.comAtprotoTempRequestPhoneVerification);
+
+/// `com.atproto.temp.transferAccount#main`
+const comAtprotoTempTransferAccount =
+    LexObject(ids.comAtprotoTempTransferAccount);
+
 /// `com.atproto.server.createSession#main`
 const comAtprotoServerCreateSession =
     LexObject(ids.comAtprotoServerCreateSession);
@@ -323,6 +331,18 @@ const comAtprotoServerDescribeServerLinks =
 /// `com.atproto.server.confirmEmail#main`
 const comAtprotoServerConfirmEmail =
     LexObject(ids.comAtprotoServerConfirmEmail);
+
+/// `com.atproto.admin.deleteCommunicationTemplate#main`
+const comAtprotoAdminDeleteCommunicationTemplate =
+    LexObject(ids.comAtprotoAdminDeleteCommunicationTemplate);
+
+/// `com.atproto.admin.updateCommunicationTemplate#main`
+const comAtprotoAdminUpdateCommunicationTemplate =
+    LexObject(ids.comAtprotoAdminUpdateCommunicationTemplate);
+
+/// `com.atproto.admin.deleteAccount#main`
+const comAtprotoAdminDeleteAccount =
+    LexObject(ids.comAtprotoAdminDeleteAccount);
 
 /// `com.atproto.server.getSession#main`
 const comAtprotoServerGetSession = LexObject(ids.comAtprotoServerGetSession);
@@ -443,6 +463,22 @@ const comAtprotoRepoPutRecord = LexObject(ids.comAtprotoRepoPutRecord);
 /// `com.atproto.repo.uploadBlob#main`
 const comAtprotoRepoUploadBlob = LexObject(ids.comAtprotoRepoUploadBlob);
 
+/// `com.atproto.admin.defs#communicationTemplateView`
+const comAtprotoAdminDefsCommunicationTemplateView =
+    LexObject(ids.comAtprotoAdminDefsCommunicationTemplateView);
+
+/// `com.atproto.admin.queryModerationEvents#main`
+const comAtprotoAdminQueryModerationEvents =
+    LexObject(ids.comAtprotoAdminQueryModerationEvents);
+
+/// `com.atproto.admin.createCommunicationTemplate#main`
+const comAtprotoAdminCreateCommunicationTemplate =
+    LexObject(ids.comAtprotoAdminCreateCommunicationTemplate);
+
+/// `com.atproto.admin.disableInviteCodes#main`
+const comAtprotoAdminDisableInviteCodes =
+    LexObject(ids.comAtprotoAdminDisableInviteCodes);
+
 /// `com.atproto.repo.describeRepo#main`
 const comAtprotoRepoDescribeRepo = LexObject(ids.comAtprotoRepoDescribeRepo);
 
@@ -470,6 +506,14 @@ const comAtprotoRepoListRecords = LexObject(ids.comAtprotoRepoListRecords);
 /// `com.atproto.repo.listRecords#record`
 const comAtprotoRepoListRecordsRecord =
     LexObject(ids.comAtprotoRepoListRecordsRecord);
+
+/// `com.atproto.admin.listCommunicationTemplates#main`
+const comAtprotoAdminListCommunicationTemplates =
+    LexObject(ids.comAtprotoAdminListCommunicationTemplates);
+
+/// `com.atproto.admin.disableAccountInvites#main`
+const comAtprotoAdminDisableAccountInvites =
+    LexObject(ids.comAtprotoAdminDisableAccountInvites);
 
 /// `com.atproto.moderation.createReport#main`
 const comAtprotoModerationCreateReport =