+++ packages/atproto/lib/src/services/sync_service.dart
@@ -239,7 +239,7 @@ final class SyncService {
     required String did,
     required String? sinceCommitCid,
     required core.ProgressStatus? progress,
-    core.To<T>? to,
+    core.ResponseDataBuilder<T>? to,
   }) async =>
       await _ctx.get(
         ns.comAtprotoSyncGetRepo,
@@ -257,7 +257,7 @@ final class SyncService {
   Future<core.XRPCResponse<T>> _findRepoBlocks<T>({
     required String did,
     required List<String> commitCids,
-    core.To<T>? to,
+    core.ResponseDataBuilder<T>? to,
   }) async =>
       await _ctx.get(
         ns.comAtprotoSyncGetBlocks,
@@ -271,7 +271,7 @@ final class SyncService {
 
   Future<core.XRPCResponse<T>> _findLatestCommit<T>({
     required String did,
-    core.To<T>? to,
+    core.ResponseDataBuilder<T>? to,
   }) async =>
       await _ctx.get(
         ns.comAtprotoSyncGetLatestCommit,
@@ -284,7 +284,7 @@ final class SyncService {
   Future<core.XRPCResponse<T>> _findRecord<T>({
     required core.AtUri uri,
     required String? commitCid,
-    core.To<T>? to,
+    core.ResponseDataBuilder<T>? to,
   }) async =>
       await _ctx.get(
         ns.comAtprotoSyncGetRecord,
@@ -301,7 +301,7 @@ final class SyncService {
   Future<core.XRPCResponse<T>> _findRepos<T>({
     required int? limit,
     required String? cursor,
-    core.To<T>? to,
+    core.ResponseDataBuilder<T>? to,
   }) async =>
       await _ctx.get(
         ns.comAtprotoSyncListRepos,
@@ -317,7 +317,7 @@ final class SyncService {
     required String? sinceCid,
     required int? limit,
     required String? cursor,
-    core.To<T>? to,
+    core.ResponseDataBuilder<T>? to,
   }) async =>
       await _ctx.get(
         ns.comAtprotoSyncListBlobs,