+++ packages/bluesky/lib/bluesky.dart
@@ -195,6 +195,8 @@ export 'package:atproto_core/atproto_core.dart'
         HttpStatus,
         Serializable,
         Platform,
+        ResponseDataBuilder,
+        ResponseDataAdaptor,
         AtUri,
         NSID,
         CID,