+++ packages/bluesky/lib/lex_namespaces.dart
@@ -1,0 +1,6 @@
+// Copyright 2024 Shinya Kato. All rights reserved.
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided the conditions.
+
+export 'package:atproto/lex_namespaces.dart';
+export 'package:bluesky/src/nsids.g.dart';