+++ packages/bluesky/lib/src/services/actor_service.dart
@@ -203,7 +203,7 @@ final class ActorService {
     required String term,
     required int? limit,
     required String? cursor,
-    core.To<T>? to,
+    core.ResponseDataBuilder<T>? to,
   }) async =>
       await _ctx.get(
         ns.appBskyActorSearchActors,
@@ -217,7 +217,7 @@ final class ActorService {
 
   Future<core.XRPCResponse<T>> _findProfile<T>({
     required String actor,
-    core.To<T>? to,
+    core.ResponseDataBuilder<T>? to,
   }) async =>
       await _ctx.get(
         ns.appBskyActorGetProfile,
@@ -229,7 +229,7 @@ final class ActorService {
 
   Future<core.XRPCResponse<T>> _findProfiles<T>({
     required List<String> actors,
-    core.To<T>? to,
+    core.ResponseDataBuilder<T>? to,
   }) async =>
       await _ctx.get(
         ns.appBskyActorGetProfiles,
@@ -242,7 +242,7 @@ final class ActorService {
   Future<core.XRPCResponse<T>> _findSuggestions<T>({
     required int? limit,
     required String? cursor,
-    core.To<T>? to,
+    core.ResponseDataBuilder<T>? to,
   }) async =>
       await _ctx.get(
         ns.appBskyActorGetSuggestions,
@@ -256,7 +256,7 @@ final class ActorService {
   Future<core.XRPCResponse<T>> _searchTypeahead<T>({
     required String term,
     required int? limit,
-    core.To<T>? to,
+    core.ResponseDataBuilder<T>? to,
   }) async =>
       await _ctx.get(
         ns.appBskyActorSearchActorsTypeahead,
@@ -268,7 +268,7 @@ final class ActorService {
       );
 
   Future<core.XRPCResponse<T>> _findPreferences<T>({
-    core.To<T>? to,
+    core.ResponseDataBuilder<T>? to,
   }) async =>
       await _ctx.get(
         ns.appBskyActorGetPreferences,