+++ packages/bluesky/lib/src/services/notification_service.dart
@@ -110,7 +110,7 @@ final class NotificationService {
     required int? limit,
     required String? cursor,
     required DateTime? seenAt,
-    core.To<T>? to,
+    core.ResponseDataBuilder<T>? to,
   }) async =>
       await _ctx.get(
         ns.appBskyNotificationListNotifications,
@@ -123,7 +123,7 @@ final class NotificationService {
       );
 
   Future<core.XRPCResponse<T>> _findUnreadCount<T>({
-    core.To<T>? to,
+    core.ResponseDataBuilder<T>? to,
   }) async =>
       await _ctx.get(
         ns.appBskyNotificationGetUnreadCount,