+++ packages/bluesky/lib/src/services/unspecced_service.dart
@@ -135,7 +135,7 @@ final class UnspeccedService {
     required int? limit,
     required String? cursor,
     required String? query,
-    core.To<T>? to,
+    core.ResponseDataBuilder<T>? to,
   }) async =>
       await _ctx.get(
         ns.appBskyUnspeccedGetPopularFeedGenerators,
@@ -150,7 +150,7 @@ final class UnspeccedService {
   Future<core.XRPCResponse<T>> _findTimelineSkeleton<T>({
     required int? limit,
     required String? cursor,
-    core.To<T>? to,
+    core.ResponseDataBuilder<T>? to,
   }) async =>
       await _ctx.get(
         ns.appBskyUnspeccedGetTimelineSkeleton,
@@ -165,7 +165,7 @@ final class UnspeccedService {
     required String query,
     required int? limit,
     required String? cursor,
-    core.To<T>? to,
+    core.ResponseDataBuilder<T>? to,
   }) async =>
       await _ctx.get(
         ns.appBskyUnspeccedSearchPostsSkeleton,
@@ -182,7 +182,7 @@ final class UnspeccedService {
     required bool? typeahead,
     required int? limit,
     required String? cursor,
-    core.To<T>? to,
+    core.ResponseDataBuilder<T>? to,
   }) async =>
       await _ctx.get(
         ns.appBskyUnspeccedSearchActorsSkeleton,