+++ packages/atproto/CHANGELOG.md
@@ -8,6 +8,11 @@
 
 - All objects that do not match Lexicon names have been deprecated and those that do match Lexicon names have been exposed as the `lex_types` package. You can import it like `import 'package:atproto/lex_types.dart';` ([#1164](https://github.com/myConsciousness/atproto.dart/issues/1164))
 
+## v0.11.3
+
+- Added `.get` and `.post` methods on `ATProto` object. ([#1217](https://github.com/myConsciousness/atproto.dart/issues/1217))
+- Added `atproto/lex_namespaces` package. You can use these constants for `.get` and `.post` methods.
+
 ## v0.11.2
 
 - Upgraded `atproto_core`. ([#1159](https://github.com/myConsciousness/atproto.dart/issues/1159))